@@ -3,7 +3,7 @@
   "version": "1.0.0",
   "description": "Backend starter template",
   "main": "dist/server.js",
-  "repository": "https://github.com/backmeupplz/backend-starter",
+  "repository": "https://github.com/Borodutch/backend-starter",
   "author": "backmeupplz <backmeupplz@gmail.com>",
   "license": "MIT",
   "private": false,
@@ -13,40 +13,48 @@
     "build-ts": "tsc --skipLibCheck",
     "watch-ts": "tsc -w --skipLibCheck",
     "watch-js": "nodemon dist/server.js",
-    "test": "yarn jest --forceExit"
-  },
-  "devDependencies": {
-    "@koa/cors": "^3.1.0",
-    "@typegoose/typegoose": "^7.6.1",
-    "@types/dotenv": "^8.2.0",
-    "@types/jest": "^26.0.20",
-    "@types/jsonwebtoken": "^8.5.2",
-    "@types/koa": "^2.13.3",
-    "@types/lodash": "^4.14.168",
-    "@types/mongoose": "^5.10.2",
-    "@types/node": "^15.12.5",
-    "concurrently": "^6.0.0",
-    "jest": "^27.0.6",
-    "mongodb-memory-server": "^6.9.6",
-    "nodemon": "^2.0.9",
-    "supertest": "^6.1.3",
-    "ts-jest": "^27.0.3",
-    "typescript": "^4.3.5"
+    "test": "yarn jest --forceExit",
+    "pretty": "prettier --check src",
+    "lint": "yarn pretty && eslint --max-warnings 0 --ext ts,tsx,json src"
   },
   "dependencies": {
-    "@types/koa-bodyparser": "^4.3.1",
-    "amala": "^5.1.2",
-    "axios": "^0.21.1",
-    "axios-mock-adapter": "^1.19.0",
-    "dotenv": "^10.0.0",
+    "@hapi/boom": "^9.1.4",
+    "@koa/cors": "^3.1.0",
+    "amala": "^7.0.0",
+    "axios": "^0.22.0",
     "facebook-node-sdk": "^0.2.0",
     "jsonwebtoken": "^8.5.1",
-    "koa": "^2.13.1",
+    "koa": "^2.13.3",
     "koa-bodyparser": "^4.3.0",
-    "koa-router": "^10.0.0",
+    "koa-router": "^10.1.1",
     "lodash": "^4.17.21",
+    "mongoose": "6.0.10"
+  },
+  "devDependencies": {
+    "@typegoose/typegoose": "^9.1.0",
+    "@types/jest": "^27.0.2",
+    "@types/jsonwebtoken": "^8.5.5",
+    "@types/koa": "^2.13.4",
+    "@types/koa-bodyparser": "^4.3.3",
+    "@types/lodash": "^4.14.175",
+    "@types/node": "^16.10.3",
+    "@typescript-eslint/eslint-plugin": "^4.33.0",
+    "@typescript-eslint/parser": "^4.33.0",
+    "axios-mock-adapter": "^1.20.0",
+    "concurrently": "^6.3.0",
+    "dotenv": "^10.0.0",
+    "eslint": "^7.32.0",
+    "eslint-config-prettier": "^8.3.0",
+    "eslint-plugin-prettier": "^4.0.0",
+    "eslint-plugin-sort-imports-es6-autofix": "^0.6.0",
+    "jest": "^27.2.5",
     "module-alias": "^2.2.2",
-    "mongoose": "5.10.18"
+    "mongodb-memory-server": "^7.4.3",
+    "nodemon": "^2.0.13",
+    "prettier": "^2.4.1",
+    "supertest": "^6.1.6",
+    "ts-jest": "^27.0.5",
+    "typescript": "^4.4.3"
   },
   "_moduleAliases": {
     "@": "dist"
