@@ -1,18 +1,20 @@
-import axios from 'axios'
-import { Context } from 'koa'
-import { getOrCreateUser } from '@/models/user'
+import * as Facebook from 'facebook-node-sdk'
 import { Body, Controller, Ctx, Post } from 'amala'
-import Facebook = require('facebook-node-sdk')
+import { Context } from 'koa'
 import {
   TelegramLoginPayload,
   verifyTelegramPayload,
 } from '@/helpers/verifyTelegramPayload'
+import { forbidden } from '@hapi/boom'
+import { getOrCreateUser } from '@/models/user'
+import FBUser from '@/models/FBUser'
+import axios from 'axios'
 
 @Controller('/login')
 export default class LoginController {
   @Post('/facebook')
   async facebook(@Body('accessToken') accessToken: string) {
-    const fbProfile: any = await getFBUser(accessToken)
+    const fbProfile: FBUser = await getFBUser(accessToken)
     const user = await getOrCreateUser({
       name: fbProfile.name,
 
@@ -26,7 +28,7 @@ export default class LoginController {
   async telegram(@Ctx() ctx: Context, @Body() data: TelegramLoginPayload) {
     // verify the data
     if (!verifyTelegramPayload(data)) {
-      return ctx.throw(403)
+      return ctx.throw(forbidden())
     }
 
     const user = await getOrCreateUser({
@@ -38,7 +40,7 @@ export default class LoginController {
 
   @Post('/google')
   async google(@Body('accessToken') accessToken: string) {
-    const userData: any =
+    const userData =
       process.env.TESTING === 'true'
         ? testingGoogleMock()
         : (
@@ -62,7 +64,7 @@ export default class LoginController {
   }
 }
 
-function getFBUser(accessToken: string) {
+function getFBUser(accessToken: string): Promise<FBUser> {
   return new Promise((res, rej) => {
     const fb = new Facebook({
       appID: process.env.FACEBOOK_APP_ID,
