@@ -1,9 +1,9 @@
 import 'reflect-metadata'
 import * as Koa from 'koa'
-import * as bodyParser from 'koa-bodyparser'
-import { bootstrapControllers } from 'amala'
-import * as cors from '@koa/cors'
 import * as Router from 'koa-router'
+import * as bodyParser from 'koa-bodyparser'
+import * as cors from '@koa/cors'
+import { bootstrapControllers } from 'amala'
 
 export const app = new Koa()
 ;(async () => {
