@@ -2,7 +2,7 @@ import * as jwt from 'jsonwebtoken'
 
 const secret = process.env.JWT
 
-export function sign(payload: object) {
+export function sign(payload: Record<string, unknown>) {
   return new Promise((res, rej) => {
     jwt.sign(payload, secret, undefined, (err, token) => {
       return err ? rej(err) : res(token)
